functions/main-coordinator/requirements.txt
functions-framework==3.*
google-cloud-logging==3.*
google-adk